Start migrating the appconfig-related endpoints
