+++ server/service/handler.go
@@ -29,10 +29,6 @@ type FleetEndpoints struct {
 	ResetPassword                         endpoint.Endpoint
 	CreateUserWithInvite                  endpoint.Endpoint
 	PerformRequiredPasswordReset          endpoint.Endpoint
-	GetAppConfig                          endpoint.Endpoint
-	ModifyAppConfig                       endpoint.Endpoint
-	ApplyEnrollSecretSpec                 endpoint.Endpoint
-	GetEnrollSecretSpec                   endpoint.Endpoint
 	CreateInvite                          endpoint.Endpoint
 	ListInvites                           endpoint.Endpoint
 	DeleteInvite                          endpoint.Endpoint
@@ -57,14 +53,12 @@ type FleetEndpoints struct {
 	CarveBegin                            endpoint.Endpoint
 	CarveBlock                            endpoint.Endpoint
 	SearchTargets                         endpoint.Endpoint
-	GetCertificate                        endpoint.Endpoint
 	ChangeEmail                           endpoint.Endpoint
 	InitiateSSO                           endpoint.Endpoint
 	CallbackSSO                           endpoint.Endpoint
 	SSOSettings                           endpoint.Endpoint
 	StatusResultStore                     endpoint.Endpoint
 	StatusLiveQuery                       endpoint.Endpoint
-	Version                               endpoint.Endpoint
 }
 
 // MakeFleetServerEndpoints creates the Fleet API endpoints.
@@ -93,11 +87,6 @@ func MakeFleetServerEndpoints(svc fleet.Service, urlPrefix string, limitStore th
 		PerformRequiredPasswordReset: logged(canPerformPasswordReset(makePerformRequiredPasswordResetEndpoint(svc))),
 
 		// Standard user authentication routes
-		GetAppConfig:          authenticatedUser(svc, makeGetAppConfigEndpoint(svc)),
-		ModifyAppConfig:       authenticatedUser(svc, makeModifyAppConfigEndpoint(svc)),
-		ApplyEnrollSecretSpec: authenticatedUser(svc, makeApplyEnrollSecretSpecEndpoint(svc)),
-		GetEnrollSecretSpec:   authenticatedUser(svc, makeGetEnrollSecretSpecEndpoint(svc)),
-
 		CreateInvite: authenticatedUser(svc, makeCreateInviteEndpoint(svc)),
 		ListInvites:  authenticatedUser(svc, makeListInvitesEndpoint(svc)),
 		DeleteInvite: authenticatedUser(svc, makeDeleteInviteEndpoint(svc)),
@@ -115,9 +104,7 @@ func MakeFleetServerEndpoints(svc fleet.Service, urlPrefix string, limitStore th
 		CreateDistributedQueryCampaign:        authenticatedUser(svc, makeCreateDistributedQueryCampaignEndpoint(svc)),
 		CreateDistributedQueryCampaignByNames: authenticatedUser(svc, makeCreateDistributedQueryCampaignByNamesEndpoint(svc)),
 		SearchTargets:                         authenticatedUser(svc, makeSearchTargetsEndpoint(svc)),
-		GetCertificate:                        authenticatedUser(svc, makeCertificateEndpoint(svc)),
 		ChangeEmail:                           authenticatedUser(svc, makeChangeEmailEndpoint(svc)),
-		Version:                               authenticatedUser(svc, makeVersionEndpoint(svc)),
 
 		// Authenticated status endpoints
 		StatusResultStore: authenticatedUser(svc, makeStatusResultStoreEndpoint(svc)),
@@ -145,10 +132,6 @@ type fleetHandlers struct {
 	ResetPassword                         http.Handler
 	CreateUserWithInvite                  http.Handler
 	PerformRequiredPasswordReset          http.Handler
-	GetAppConfig                          http.Handler
-	ModifyAppConfig                       http.Handler
-	ApplyEnrollSecretSpec                 http.Handler
-	GetEnrollSecretSpec                   http.Handler
 	CreateInvite                          http.Handler
 	ListInvites                           http.Handler
 	DeleteInvite                          http.Handler
@@ -173,14 +156,12 @@ type fleetHandlers struct {
 	CarveBegin                            http.Handler
 	CarveBlock                            http.Handler
 	SearchTargets                         http.Handler
-	GetCertificate                        http.Handler
 	ChangeEmail                           http.Handler
 	InitiateSSO                           http.Handler
 	CallbackSSO                           http.Handler
 	SettingsSSO                           http.Handler
 	StatusResultStore                     http.Handler
 	StatusLiveQuery                       http.Handler
-	Version                               http.Handler
 }
 
 func makeKitHandlers(e FleetEndpoints, opts []kithttp.ServerOption) *fleetHandlers {
@@ -195,10 +176,6 @@ func makeKitHandlers(e FleetEndpoints, opts []kithttp.ServerOption) *fleetHandle
 		ResetPassword:                         newServer(e.ResetPassword, decodeResetPasswordRequest),
 		CreateUserWithInvite:                  newServer(e.CreateUserWithInvite, decodeCreateUserRequest),
 		PerformRequiredPasswordReset:          newServer(e.PerformRequiredPasswordReset, decodePerformRequiredPasswordResetRequest),
-		GetAppConfig:                          newServer(e.GetAppConfig, decodeNoParamsRequest),
-		ModifyAppConfig:                       newServer(e.ModifyAppConfig, decodeModifyAppConfigRequest),
-		ApplyEnrollSecretSpec:                 newServer(e.ApplyEnrollSecretSpec, decodeApplyEnrollSecretSpecRequest),
-		GetEnrollSecretSpec:                   newServer(e.GetEnrollSecretSpec, decodeNoParamsRequest),
 		CreateInvite:                          newServer(e.CreateInvite, decodeCreateInviteRequest),
 		ListInvites:                           newServer(e.ListInvites, decodeListInvitesRequest),
 		DeleteInvite:                          newServer(e.DeleteInvite, decodeDeleteInviteRequest),
@@ -223,14 +200,12 @@ func makeKitHandlers(e FleetEndpoints, opts []kithttp.ServerOption) *fleetHandle
 		CarveBegin:                            newServer(e.CarveBegin, decodeCarveBeginRequest),
 		CarveBlock:                            newServer(e.CarveBlock, decodeCarveBlockRequest),
 		SearchTargets:                         newServer(e.SearchTargets, decodeSearchTargetsRequest),
-		GetCertificate:                        newServer(e.GetCertificate, decodeNoParamsRequest),
 		ChangeEmail:                           newServer(e.ChangeEmail, decodeChangeEmailRequest),
 		InitiateSSO:                           newServer(e.InitiateSSO, decodeInitiateSSORequest),
 		CallbackSSO:                           newServer(e.CallbackSSO, decodeCallbackSSORequest),
 		SettingsSSO:                           newServer(e.SSOSettings, decodeNoParamsRequest),
 		StatusResultStore:                     newServer(e.StatusResultStore, decodeNoParamsRequest),
 		StatusLiveQuery:                       newServer(e.StatusLiveQuery, decodeNoParamsRequest),
-		Version:                               newServer(e.Version, decodeNoParamsRequest),
 	}
 }
 
@@ -412,11 +387,6 @@ func attachFleetAPIRoutes(r *mux.Router, h *fleetHandlers) {
 
 	r.Handle("/api/v1/fleet/users", h.CreateUserWithInvite).Methods("POST").Name("create_user_with_invite")
 
-	r.Handle("/api/v1/fleet/config/certificate", h.GetCertificate).Methods("GET").Name("get_certificate")
-	r.Handle("/api/v1/fleet/config", h.GetAppConfig).Methods("GET").Name("get_app_config")
-	r.Handle("/api/v1/fleet/config", h.ModifyAppConfig).Methods("PATCH").Name("modify_app_config")
-	r.Handle("/api/v1/fleet/spec/enroll_secret", h.ApplyEnrollSecretSpec).Methods("POST").Name("apply_enroll_secret_spec")
-	r.Handle("/api/v1/fleet/spec/enroll_secret", h.GetEnrollSecretSpec).Methods("GET").Name("get_enroll_secret_spec")
 	r.Handle("/api/v1/fleet/invites", h.CreateInvite).Methods("POST").Name("create_invite")
 	r.Handle("/api/v1/fleet/invites", h.ListInvites).Methods("GET").Name("list_invites")
 	r.Handle("/api/v1/fleet/invites/{id:[0-9]+}", h.DeleteInvite).Methods("DELETE").Name("delete_invite")
@@ -439,8 +409,6 @@ func attachFleetAPIRoutes(r *mux.Router, h *fleetHandlers) {
 
 	r.Handle("/api/v1/fleet/targets", h.SearchTargets).Methods("POST").Name("search_targets")
 
-	r.Handle("/api/v1/fleet/version", h.Version).Methods("GET").Name("version")
-
 	r.Handle("/api/v1/fleet/status/result_store", h.StatusResultStore).Methods("GET").Name("status_result_store")
 	r.Handle("/api/v1/fleet/status/live_query", h.StatusLiveQuery).Methods("GET").Name("status_live_query")
 
@@ -459,6 +427,13 @@ func attachNewStyleFleetAPIRoutes(r *mux.Router, svc fleet.Service, opts []kitht
 	e.GET("/api/_version_/fleet/me", meEndpoint, nil)
 	e.GET("/api/_version_/fleet/sessions/{id:[0-9]+}", getInfoAboutSessionEndpoint, getInfoAboutSessionRequest{})
 	e.DELETE("/api/_version_/fleet/sessions/{id:[0-9]+}", deleteSessionEndpoint, deleteSessionRequest{})
+
+	e.GET("/api/_version_/fleet/config/certificate", getCertificateEndpoint, nil)
+	e.GET("/api/_version_/fleet/config", getAppConfigEndpoint, nil)
+	e.PATCH("/api/_version_/fleet/config", modifyAppConfigEndpoint, modifyAppConfigRequest{})
+	e.POST("/api/_version_/fleet/spec/enroll_secret", applyEnrollSecretSpecEndpoint, applyEnrollSecretSpecRequest{})
+	e.GET("/api/_version_/fleet/spec/enroll_secret", getEnrollSecretSpecEndpoint, nil)
+	e.GET("/api/_version_/fleet/version", versionEndpoint, nil)
 
 	e.POST("/api/_version_/fleet/users/roles/spec", applyUserRoleSpecsEndpoint, applyUserRoleSpecsRequest{})
 	e.POST("/api/_version_/fleet/translate", translatorEndpoint, translatorRequest{})
