+++ server/service/handler_test.go
@@ -56,14 +56,6 @@ func TestAPIRoutes(t *testing.T) {
 		},
 		{
 			verb: "GET",
-			uri:  "/api/v1/fleet/config",
-		},
-		{
-			verb: "PATCH",
-			uri:  "/api/v1/fleet/config",
-		},
-		{
-			verb: "GET",
 			uri:  "/api/v1/fleet/invites",
 		},
 		{
